Use Tuist instead of CMake as a project generator
Tuist provides a more "native" experience by building Xcode projects based on a
configuration file. Even though CMake also technically supports Xcode, it is
much more complex to set it up properly.

Since the project is not cross-platform and Xcode already provides a lot of
developer and debugging tools out of the box, it makes sense to use it as the
primary development environment.