Fix frames processing after capture stop
Even after the capture being stopped, some frames were still being processed.
That was happening because the capture happens in a different thread, and then
it schedules a callback to the main thread to process the frame.

The error was causing an app crash on some devices, specifically when user
switched between different resolution frames (because the app tried to find a
difference between e.g. 1280x720 and 1920x1080 images).

In order to prevent frames being processed after the capture is stopped, capture
session id is introduced. If it changes at the moment of main thread dispatch,
it means the frames should be ignored as the capture session changed.