Set target bitrate for video capture
Video capture was using default bitrate, which was usually consuming too much
space. Telegram bot API only allows sending a video up to 50MB, so we need to
set a proper bitrate.

The bitrate will be set to 300 KB/s, which should result in a bit less than 3
minute video of size 50MB.