Use UserDefaults instead of secure storage for telegram api token
Since the app is not signed and not plannnig to be signed, secure storage (that
uses keychain under the hood) will block access to the data after each new app
update. It will prompt the user to explicitly allow access and there is no way
to avoid that. Furthermore, there is no control over the prompt, e.g. the OS
will not inform the app about it, only return the result.

UserDefaults is not encrypted, but there is no other option at the moment.